fix(compiler): split compiler entry points
Move module and function compilation into separate files and shift function memory declaration validation into the tokenizer.
