+++ packages/compiler/tests/compilerOptionsAndCache.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, test } from 'vitest';
 import { compileFixtureProgramSource } from './testUtils';
 
 describe('compiler options and cache', () => {
-	test('exposes AST and stack analysis only when requested', () => {
+	test('always exposes AST and exposes stack analysis only when requested', () => {
 		const source = `
 8f4e/v1
 
@@ -25,15 +25,14 @@ functionEnd int
 `;
 		const defaultResult = compileFixtureProgramSource(source).compileResult;
 		const analyzedResult = compileFixtureProgramSource(source, {
-			includeAST: true,
 			includeStackAnalysis: true,
 		}).compileResult;
 
-		expect(defaultResult.compiledModules.optionSurface.ast).toBeUndefined();
+		expect(defaultResult.compiledModules.optionSurface.ast.id).toBe('optionSurface');
 		expect(defaultResult.compiledModules.optionSurface.stackAnalysis).toBeUndefined();
 		expect(defaultResult.compiledFunctions!.double.stackAnalysis).toBeUndefined();
 
-		expect(analyzedResult.compiledModules.optionSurface.ast?.id).toBe('optionSurface');
+		expect(analyzedResult.compiledModules.optionSurface.ast.id).toBe('optionSurface');
 		expect(analyzedResult.compiledModules.optionSurface.stackAnalysis?.map(line => line.instruction)).toEqual([
 			'push',
 			'push',
@@ -63,11 +62,10 @@ function cachedFunction
 push 1
 functionEnd int
 `;
-		const first = compileFixtureProgramSource(source, { includeAST: true });
+		const first = compileFixtureProgramSource(source);
 		const cachedModuleAst = first.compileResult.cache.ast.entries.get('entry:main:module:0')?.ast;
 		const cachedFunctionAst = first.compileResult.cache.ast.entries.get('function:0')?.ast;
 		const second = compileFixtureProgramSource(source, {
-			includeAST: true,
 			cache: first.compileResult.cache,
 		});
